Replaced use of CMake path variables
to use PROJECT_SOURCE_DIR and not CMAKE_CURRENT_SOURCE_DIR when referring to
subdirectories of the toplevel dir. This prevents problems when the top-level
CMakeLists is included in another CMake project.

The problem arises because CMAKE_CURRENT_SOURCE_DIR always refers to the
currently parsed CMakeLists.txt, which might be the wrong one if it is
included somewhere else