better check for variadic template support
The previous check failed to detect problems with inclomplete C++11 support:
It checked only for variadic template support, the problem aries,
however, not only because of lack of variadic templates but because of lack of
support for template aliases. Since compilers that support template aliases
should always also support variadic templates, the check is not for that.

The two checks are simply to support multiple boost versions.